src/trading/requirements.txt
# 基本パッケージ
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0

# データ取得
yfinance>=0.1.70
alpha_vantage>=2.3.1
pandas-datareader>=0.10.0

# 取引実行
ib_insync>=0.9.70

# 環境変数
python-dotenv>=0.19.0

# 非同期処理
aiohttp>=3.8.0
asyncio>=3.4.3

# テクニカル分析
ta>=0.10.0
ta-lib>=0.4.24

# 機械学習（オプション）
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0